Use CTA from capi in hosted content
