+++ common/app/common/commercial/hosted/HostedArticlePage.scala
@@ -72,6 +72,9 @@ object HostedArticlePage extends Logging {
       sponsorships <- hostedTag.activeSponsorships
       sponsorship <- sponsorships.headOption
       toneTag <- tags find (_.`type` == TagType.Tone)
+      atoms <- content.atoms
+      ctaAtoms <- atoms.cta
+      ctaAtom <- ctaAtoms.headOption
     } yield {
 
       val mainImageAsset: Option[Asset] = {
@@ -102,19 +105,11 @@ object HostedArticlePage extends Logging {
         // using capi trail text instead of standfirst because we don't want the markup
         standfirst = content.fields.flatMap(_.trailText).getOrElse(""),
         body = content.fields.flatMap(_.body).getOrElse(""),
-        // todo: from cta atom
-        cta = HostedCallToAction(
-          url = "http://www.actforwildlife.org.uk/?utm_source=theguardian.com&utm_medium=referral&utm_campaign=LaunchCampaignSep2016",
-          image = Some("http://media.guim.co.uk/d723e82cdd399f013905a5ee806fea3591b4a363/0_926_3872_1666/2000.jpg"),
-          label = Some("It's time to act for wildlife"),
-          trackingCode = Some("act-for-wildlife-button"),
-          btnText = Some("Act for wildlife")
-        ),
+        cta = HostedCallToAction.fromAtom(ctaAtom),
         mainPicture = mainImageAsset.flatMap(_.file) getOrElse "",
         mainPictureCaption = mainImageAsset.flatMap(_.typeData.flatMap(_.caption)).getOrElse(""),
         socialShareText = content.fields.flatMap(_.socialShareText),
         shortSocialShareText = content.fields.flatMap(_.shortSocialShareText),
-        // todo: related content
         nextPagesList = HostedPages.nextPages(campaignName = campaignId, pageName = content.webUrl.split(campaignId + "/")(1))
       )
     }
